Update README to show proper `mongoc:connect/3` args
Resolves #206

The README has been showing confusing information for `mongoc`, namely
suggesting to connect through `mongo_api:connect/4` but describing the
`Seed` parameter from `mongoc` as if it were part of the `mongo_api`
method, which it is not.

In this patch I've updated the README so that it describes the use of
`mongoc:connect/3` instead of `mongo_api:connect/4` as hinted at by
the descriptive text above the example.

Additionally I have made small edits to the `mongoc` area:
 - Fix a couple simple typos
 - Surround the code blocks in code fences specifying Erlang
   in order to get proper highlighting in GitHub
 - Remove references to the `Type` and `Hosts` args from `mongo_api:connect/4`
   which are otherwise not directly discussed